feat: consume providers + content-model from kbexplorer-engine package (slice 2/5)
Point src/engine/{providers.ts,providers/*,plugin-loader.ts,content-model/*}
at @anokye-labs/kbexplorer-engine (pinned to git+...#23760f381e4e2e15bc37aaa0ee9a3f336fb9ee6d,
slice 2's merged SHA, pending real npm publish).

Superseded tests deleted (now covered by engine's own green suite):
providers/*.test.ts, provider-identity-shapes.test.ts, providers.test.ts,
content-model/__tests__/{builder,schema-reader,service-decision,vocabulary,
work-graph}.test.ts, providers/rich-markdown/__tests__/detect.test.ts.

Kept / restored (real cross-slice dependencies found during verification,
not superseded by the move):
- content-model/__tests__/fixtures.ts (+ its fixture tree) - still used by
  template-local pipeline-idempotency.test.ts and source-edit.test.ts, which
  build a ContentModelSource and feed it through the package's
  buildContentModel. Updated its ContentModelSource import to the package.
- providers/rich-markdown/shims/{crypto,path}.ts + shims.test.ts - these are
  browser polyfills aliased at the Vite-bundler level in vite.config.ts
  (node:crypto / node:path -> these files) for the whole app build, entirely
  independent of whether the provider itself is a shim now.
- providers/examples/glossary-provider.ts - a self-contained
  defineProvider() example referenced by docs/providers.md as the canonical
  local-provider (F5a) reference; kept as documentation/reference material.

New: views/rich-markdown/__tests__/authored-rich-markdown-provider-integration.test.ts
restores, as a template-side cross-layer integration test, the 3 fully-dropped
+ 2 partially-trimmed assertions from engine's authored-rich-markdown-provider
test that depended on template's views/rich-markdown (isRichMarkdownNode,
getRichMarkdownDocument, planProseFence) - see kbexplorer-engine#3 for the
judgment-call writeup.

Verified: build (tsc -b + vite build) green, 733/733 tests green (73 files),
lint clean (0 errors, 3 pre-existing unrelated warnings).

Refs #472, #463.
